Create Password.java
This program has one value of username and one value of password. If both your inputs are correct, the user will log in. If at least one of their inputs is wrong, then the system will output "One or all your entries are wrong :(" and ask the user whether they want to change their password or not. If the user says "Yes", then they will be able to change their password.